Refactor route details to own file
