+++ src/components/map/index.ts
@@ -1,1 +1,1 @@
-export * from './Map'; // eslint-disable-line import/no-cycle
+export * from './Map';
