+++ src/components/map/stops/EditStopLayer.tsx
@@ -21,7 +21,6 @@ import {
   mapLngLatToPoint,
   showSuccessToast,
 } from '../../../utils';
-// eslint-disable-next-line import/no-cycle
 import {
   ConflictResolverModal,
   mapStopToCommonConflictItem,
