+++ src/components/map/stops/index.ts
@@ -1,5 +1,5 @@
-export * from './EditStopLayer'; // eslint-disable-line import/no-cycle
+export * from './EditStopLayer';
 export * from './EditStopModal';
 export * from './Stop';
 export * from './StopPopup';
-export * from './Stops'; // eslint-disable-line import/no-cycle
+export * from './Stops';
