+++ src/components/routes-and-lines/LinesList.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { RouteLine } from '../../generated/graphql';
-import { LineTableRow } from '../LineTableRow'; // eslint-disable-line import/no-cycle
+import { LineTableRow } from '../LineTableRow';
 import { RoutesTable } from '../RoutesTable';
 
 type Props = {
