+++ src/components/routes-and-lines/RoutesList.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { RouteRoute } from '../../generated/graphql';
 import { RoutesTable } from '../RoutesTable';
-import { RoutesTableRow } from '../RoutesTableRow'; // eslint-disable-line import/no-cycle
+import { RoutesTableRow } from '../RoutesTableRow';
 
 type Props = {
   routes?: RouteRoute[];
