+++ src/components/routes-and-lines/search/SearchResultPage.tsx
@@ -3,8 +3,8 @@ import { useTranslation } from 'react-i18next';
 import { useSearch, useSearchResults } from '../../../hooks';
 import { Container, Row } from '../../../layoutComponents';
 import { CloseIconButton } from '../../../uiComponents';
-import { LinesList } from '../LinesList'; // eslint-disable-line import/no-cycle
-import { RoutesList } from '../RoutesList'; // eslint-disable-line import/no-cycle
+import { LinesList } from '../LinesList';
+import { RoutesList } from '../RoutesList';
 import { SearchContainer } from './conditions/SearchContainer';
 import { FiltersContainer } from './filters/FiltersContainer';
