--- src/routes.ts
@@ -1,123 +1,0 @@
-import React, { FunctionComponent } from 'react';
-import { useParams } from 'react-router-dom';
-import { CreateNewLinePage } from './components/CreateNewLinePage'; // eslint-disable-line import/no-cycle
-import { Main } from './components/Main';
-import { Map } from './components/map'; // eslint-disable-line import/no-cycle
-import { EditLinePage } from './components/routes-and-lines/EditLinePage'; // eslint-disable-line import/no-cycle
-import { EditRoutePage } from './components/routes-and-lines/EditRoutePage'; // eslint-disable-line import/no-cycle
-import { LineDetailsPage } from './components/routes-and-lines/LineDetailsPage'; // eslint-disable-line import/no-cycle
-import { SearchResultPage } from './components/routes-and-lines/search/SearchResultPage'; // eslint-disable-line import/no-cycle
-import { RoutesAndLinesPage } from './components/RoutesAndLinesPage'; // eslint-disable-line import/no-cycle
-
-export enum Path {
-  root = '/',
-  routes = '/routes',
-  routesSearch = '/routes/search',
-  editRoute = '/routes/:id/edit',
-  createLine = '/lines/create',
-  lineDetails = '/lines/:id',
-  editLine = '/lines/:id/edit',
-  map = '/map',
-  exampleResource = '/example/:id',
-  fallback = '*',
-}
-
-interface Route {
-  _routerRoute: Path;
-  _exact?: boolean; // When true, will only match if the path matches the location.pathname exactly.
-  translationKey?: string;
-  getLink: (...args: any) => string; // eslint-disable-line @typescript-eslint/no-explicit-any
-  includeInNav?: boolean;
-  component: React.ComponentType;
-}
-
-const ExampleResourceRoute: FunctionComponent = () => {
-  const { id } = useParams<{ id: string }>();
-  return React.createElement(
-    'h2',
-    null,
-    `Example resource route. Resouce id: ${id}`,
-  );
-};
-
-const FallbackRoute: FunctionComponent = () => {
-  return React.createElement('p', null, `404, page not found`);
-};
-
-export const routes: Record<Path, Route> = {
-  [Path.root]: {
-    _routerRoute: Path.root,
-    _exact: true,
-    translationKey: 'routes.root',
-    getLink: () => Path.root,
-    component: Main,
-    includeInNav: true,
-  },
-  [Path.routes]: {
-    _routerRoute: Path.routes,
-    _exact: true,
-    translationKey: 'routes.routes',
-    getLink: () => Path.routes,
-    component: RoutesAndLinesPage,
-    includeInNav: true,
-  },
-  [Path.routesSearch]: {
-    _routerRoute: Path.routesSearch,
-    _exact: true,
-    translationKey: 'routes.routesSearchResult',
-    getLink: () => Path.routesSearch,
-    component: SearchResultPage,
-    includeInNav: false,
-  },
-  [Path.editRoute]: {
-    _routerRoute: Path.editRoute,
-    _exact: true,
-    translationKey: 'routes.editRoute',
-    getLink: (id: string) => Path.editRoute.replace(':id', id),
-    component: EditRoutePage,
-    includeInNav: false,
-  },
-  [Path.createLine]: {
-    _routerRoute: Path.createLine,
-    _exact: true,
-    translationKey: 'lines.createNew',
-    getLink: () => Path.createLine,
-    component: CreateNewLinePage,
-    includeInNav: false,
-  },
-  [Path.lineDetails]: {
-    _routerRoute: Path.lineDetails,
-    _exact: true,
-    translationKey: 'lines.lineDetails',
-    getLink: (id: string) => Path.lineDetails.replace(':id', id),
-    component: LineDetailsPage,
-    includeInNav: false,
-  },
-  [Path.editLine]: {
-    _routerRoute: Path.editLine,
-    _exact: true,
-    translationKey: 'lines.editLine',
-    getLink: (id: string) => Path.editLine.replace(':id', id),
-    component: EditLinePage,
-    includeInNav: false,
-  },
-  [Path.map]: {
-    _routerRoute: Path.map,
-    _exact: true,
-    translationKey: 'map.map',
-    getLink: () => Path.map,
-    // @ts-expect-error Something wrong due to forwardRef used in Map component?
-    component: Map,
-    includeInNav: true,
-  },
-  [Path.exampleResource]: {
-    _routerRoute: Path.exampleResource,
-    getLink: (id: string) => Path.exampleResource.replace(':id', id),
-    component: ExampleResourceRoute,
-  },
-  [Path.fallback]: {
-    _routerRoute: Path.fallback,
-    getLink: () => '404',
-    component: FallbackRoute,
-  },
-};
